Move backend to sub-module in gradle.
