+++ backend/build.gradle
@@ -1,0 +1,34 @@
+plugins { id 'java' }
+plugins { id 'org.springframework.boot' version '2.1.2.RELEASE' }
+plugins { id 'io.spring.dependency-management' version '1.0.6.RELEASE' }
+
+group = 'ch.zuehlke.fullstack'
+version = '0.0.1-SNAPSHOT'
+sourceCompatibility = '1.8'
+
+bootJar {
+    into('static') {
+        from('../frontend/dist/frontend/')
+    }
+}
+bootJar.dependsOn ":frontend:npm_run_dist"
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.apache.httpcomponents:httpclient'
+
+    runtimeOnly 'org.springframework.boot:spring-boot-devtools'
+
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    testImplementation 'org.springframework.security:spring-security-test'
+}
+
+// This is required to be able to execute bootRun with parameters, e.g.:
+// ./gradlew -Dspring.profiles.active=ci bootRun
+bootRun.systemProperties = System.properties
+
