+++ backend/src/main/java/ch/zuehlke/fullstack/ConnectZuehlke/apis/insight/config/InsightProperties.java
