+++ backend/src/main/java/ch/zuehlke/fullstack/ConnectZuehlke/apis/insight/service/InsightEmployeeService.java
@@ -1,6 +1,7 @@
 package ch.zuehlke.fullstack.ConnectZuehlke.apis.insight.service;
 
 import ch.zuehlke.fullstack.ConnectZuehlke.domain.Employee;
+
 import java.util.List;
 
 public interface InsightEmployeeService {
