+++ backend/src/main/resources/application-ci.properties
@@ -1,6 +1,7 @@
 # This configuration is used on the Travis CI and activated via the Environment Variables
 # Setting environment variables from .travis.yml
 # `$ export SPRING_PROFILES_ACTIVE=ci`
-insight.authentication.username=test
-insight.authentication.password=ImJustFake
-insight.url=http://localhost:8081/
+# The Insight details are not required as we are mocking things out in the CI profile.
+insight.authentication.username=
+insight.authentication.password=
+insight.url=
