+++ backend/src/test/java/ch/zuehlke/fullstack/ConnectZuehlke/ConnectZuehlkeApplicationTests.java
@@ -9,9 +9,9 @@ import org.springframework.test.context.junit4.SpringRunner;
 @SpringBootTest
 public class ConnectZuehlkeApplicationTests {
 
-	@Test
-	public void contextLoads() {
-	}
+    @Test
+    public void contextLoads() {
+    }
 
 }
