+++ build.gradle
@@ -1,25 +1,5 @@
-plugins { id 'java' }
-plugins { id 'org.springframework.boot' version '2.1.2.RELEASE' }
-plugins { id 'io.spring.dependency-management' version '1.0.6.RELEASE' }
-plugins { id 'com.moowork.node' version '1.2.0' }
-
-apply from: "frontend.gradle"
-apply from: "e2e.gradle"
-
-group = 'ch.zuehlke.fullstack'
-version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '1.8'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'org.apache.httpcomponents:httpclient'
-    runtimeOnly 'org.springframework.boot:spring-boot-devtools'
-
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation 'org.springframework.security:spring-security-test'
+allprojects {
+    wrapper {
+        gradleVersion = '5.2.1'
+    }
 }
